python-app/pyproject.toml
[tool.poetry]
name = "python-app"
version = "0.1.0"
description = "Python image processing app thingy"
authors = ["Tomas Šličius <slctomas99@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pillow = "^11.1.0"
numpy = "^2.2.2"
opencv-python = "^4.11.0.86"
torch = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
